Collect modals with same styling Create common styling for account-related modals Move overflow modals to account modals and convert to scss module
