+++ src/components/AccountModals/AccountModals.module.scss
@@ -1,6 +1,6 @@
 @use '../../assets/styles/Variables.scss';
 
-.LoginModal {
+.Modal {
     padding: 0.8rem 1.5rem 1.5rem;
     text-align: center;
     height: fit-content;
@@ -35,6 +35,6 @@
     margin-top: 1rem;
 }
 
-.GoTo {
+.BackButton {
     cursor: pointer;
 }
