+++ src/components/AccountModals/LockModal/LockModal.tsx
@@ -3,14 +3,14 @@ import { Heading3, Paragraph } from '@entur/typography'
 import { Modal } from '@entur/modal'
 import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton } from '@entur/button'
-import Check from '../../assets/images/check.png'
-import retinaCheck from '../../assets/images/check@2x.png'
-import { useUser } from '../../UserProvider'
-import { useSettings } from '../../settings/SettingsProvider'
-import { CloseButton } from '../CloseButton/CloseButton'
+import Check from '../../../assets/images/check.png'
+import retinaCheck from '../../../assets/images/check@2x.png'
+import { useUser } from '../../../UserProvider'
+import { useSettings } from '../../../settings/SettingsProvider'
+import { CloseButton } from '../../CloseButton/CloseButton'
 import { LoginModal } from '../LoginModal/LoginModal'
 import { LoginCase } from '../LoginModal/login-modal-types'
-import classes from './LockModal.module.scss'
+import classes from '../AccountModals.module.scss'
 
 interface LockModalProps {
     open: boolean
@@ -51,29 +51,25 @@ const LockModal: React.FC<LockModalProps> = ({ open, onDismiss }) => {
 
     return (
         <Modal
-            className={classes.LockModal}
+            className={classes.Modal}
             size="small"
             open={open}
             title=""
             onDismiss={onDismiss}
         >
             <CloseButton onClick={onDismiss} />
-            <div className={classes.Centered}>
-                <img
-                    className={classes.Image}
-                    src={Check}
-                    srcSet={`${retinaCheck} 2x`}
-                />
-            </div>
-            <Heading3 className={classes.Heading} margin="none">
-                Tavla er låst til din konto
-            </Heading3>
+            <img
+                src={Check}
+                srcSet={`${retinaCheck} 2x`}
+                className={classes.Image}
+            />
+            <Heading3 margin="none">Tavla er låst til din konto</Heading3>
             <Paragraph className={classes.Paragraph}>
                 Fra nå av kan bare du redigere denne tavla. Hvis du ikke vil ha
                 tavla lagret lenger, kan du slette den fra oversikten over dine
                 tavler.
             </Paragraph>
-            <GridContainer spacing="medium">
+            <GridContainer spacing="medium" className={classes.GridContainer}>
                 <GridItem small={12}>
                     <PrimaryButton
                         className={classes.SubmitButton}
