+++ src/components/AccountModals/LoginModal/EmailSent/EmailSent.tsx
@@ -3,11 +3,11 @@ import type { User } from 'firebase/auth'
 import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
-import { CloseButton } from '../../CloseButton/CloseButton'
-import Check from '../../../assets/images/check.png'
-import retinaCheck from '../../../assets/images/check@2x.png'
+import { CloseButton } from '../../../CloseButton/CloseButton'
+import Check from '../../../../assets/images/check.png'
+import retinaCheck from '../../../../assets/images/check@2x.png'
 import { ModalType } from '../login-modal-types'
-import classes from '../LoginModal.module.scss'
+import classes from '../../AccountModals.module.scss'
 
 interface EmailSentProps {
     setModalType: Dispatch<SetStateAction<ModalType>>
