+++ src/components/AccountModals/LoginModal/ResetPassword/ResetPassword.tsx
@@ -6,13 +6,13 @@ import { GridContainer, GridItem } from '@entur/grid'
 import { BackArrowIcon, EmailIcon } from '@entur/icons'
 import { PrimaryButton } from '@entur/button'
 import { Heading3, Paragraph } from '@entur/typography'
-import { auth } from '../../../UserProvider'
-import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
-import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
-import { CloseButton } from '../../CloseButton/CloseButton'
+import { auth } from '../../../../UserProvider'
+import sikkerhetBom from '../../../../assets/images/sikkerhet_bom.png'
+import retinaSikkerhetBom from '../../../../assets/images/sikkerhet_bom@2x.png'
+import { CloseButton } from '../../../CloseButton/CloseButton'
 import { ModalType } from '../login-modal-types'
-import { useFormFields } from '../../../hooks/useFormFields'
-import classes from '../LoginModal.module.scss'
+import { useFormFields } from '../../../../hooks/useFormFields'
+import classes from '../../AccountModals.module.scss'
 
 export interface UserResetPassword {
     email: string
@@ -66,7 +66,7 @@ const ResetPassword: React.FC<ResetPasswordProps> = ({
                     onClick={(): void =>
                         setModalType(ModalType.LoginEmailModal)
                     }
-                    className={classes.GoTo}
+                    className={classes.BackButton}
                 />
                 <CloseButton onClick={handleClose} />
             </div>
