+++ src/components/AccountModals/NeedToBeOwnerModal/NeedToBeOwnerModal.tsx
@@ -4,11 +4,11 @@ import { Heading3, Paragraph } from '@entur/typography'
 import { PrimaryButton } from '@entur/button'
 import { useToast } from '@entur/alert'
 import { GridContainer, GridItem } from '@entur/grid'
-import retinaSikkerhetBom from '../../assets/images/sikkerhet_bom@2x.png'
-import sikkerhetBom from '../../assets/images/sikkerhet_bom.png'
-import { useSettings } from '../../settings/SettingsProvider'
-import { CloseButton } from '../CloseButton/CloseButton'
-import './OverflowModals.scss'
+import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
+import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
+import { useSettings } from '../../../settings/SettingsProvider'
+import { CloseButton } from '../../CloseButton/CloseButton'
+import classes from '../AccountModals.module.scss'
 
 interface NeedToBeOwnerModalProps {
     open: boolean
@@ -45,19 +45,21 @@ const NeedToBeOwnerModal: React.FC<NeedToBeOwnerModalProps> = ({
             open={open}
             title=""
             onDismiss={onDismiss}
-            className="overflow-modal"
+            className={classes.Modal}
         >
             <CloseButton onClick={() => onDismiss(true)} />
-            <div className="centered">
-                <img src={sikkerhetBom} srcSet={`${retinaSikkerhetBom} 2x`} />
-            </div>
+            <img
+                src={sikkerhetBom}
+                srcSet={`${retinaSikkerhetBom} 2x`}
+                className={classes.Image}
+            />
             <Heading3 margin="none">Kun eiere kan dele en tavle</Heading3>
-            <Paragraph>
+            <Paragraph className={classes.Paragraph}>
                 For å dele tavla må du være en eier av den. Lås tavla til din
                 konto for å bli eier.
             </Paragraph>
 
-            <GridContainer spacing="medium">
+            <GridContainer spacing="medium" className={classes.GridContainer}>
                 <GridItem small={12}>
                     <PrimaryButton
                         width="fluid"
