--- src/components/LockModal/LockModal.module.scss
@@ -1,41 +1,0 @@
-@use '../../assets/styles/Variables.scss';
-
-.LockModal {
-    padding: 0.8rem 1.5rem 1.5rem;
-
-    @media (max-width: 1000px) {
-        margin: 1rem;
-        height: fit-content;
-    }
-
-    & :global {
-        .eds-grid__container {
-            padding: 10%;
-        }
-    }
-}
-
-.Centered {
-    display: flex;
-    justify-content: center;
-}
-
-.Image {
-    width: 45%;
-    margin: 0 auto;
-}
-
-.Heading {
-    text-align: center;
-}
-
-.Paragraph {
-    padding: 0 1rem;
-    color: Variables.$colors-brand-blue;
-    text-align: center;
-    margin: 0;
-}
-
-.SubmitButton {
-    margin-top: 1rem;
-}
