--- src/components/MineTavlerModal/MineTavlerModal.module.scss
@@ -1,42 +1,0 @@
-@use '../../assets/styles/Variables.scss';
-
-.MineTavlerModal {
-    padding: 0.8rem 1.5rem 1.5rem;
-
-    @media (max-width: 1000px) {
-        margin: 1rem;
-        height: fit-content;
-    }
-
-    & :global {
-        .eds-grid__container {
-            padding: 10%;
-        }
-    }
-}
-
-.Paragraph {
-    padding: 1rem 1rem 0;
-    color: Variables.$colors-brand-blue;
-    text-align: center;
-    margin: 0;
-}
-
-.Heading {
-    text-align: center;
-}
-
-.Image {
-    width: 45%;
-    margin: 0 auto;
-}
-
-.Link {
-    cursor: pointer;
-    margin-bottom: 3rem;
-}
-
-.Centered {
-    display: flex;
-    justify-content: center;
-}
