--- src/components/OverflowModals/OverflowModals.scss
@@ -1,42 +1,0 @@
-@use '../../assets/styles/Variables.scss';
-
-.overflow-modal.eds-modal__content {
-    padding: 0.8rem 1.5rem 1.5rem;
-
-    p {
-        padding: 1rem 1rem 0;
-        color: Variables.$colors-brand-blue;
-        text-align: center;
-        margin: 0;
-    }
-
-    h2,
-    h3 {
-        text-align: center;
-    }
-
-    .eds-grid__container {
-        padding: 10%;
-    }
-
-    img {
-        width: 45%;
-        margin: 0 auto;
-    }
-
-    a {
-        cursor: pointer;
-        margin-bottom: 3rem;
-    }
-
-    .centered {
-        display: flex;
-        justify-content: center;
-    }
-}
-@media (max-width: 1000px) {
-    .overflow-modal {
-        margin: 1rem;
-        height: fit-content;
-    }
-}
