+++ src/containers/Admin/LockAndViewButtons/LockAndViewButtons.tsx
@@ -2,7 +2,7 @@ import React, { useState, useCallback } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import classNames from 'classnames'
 import { useSettings } from '../../../settings/SettingsProvider'
-import { LockModal } from '../../../components/LockModal/LockModal'
+import { LockModal } from '../../../components/AccountModals/LockModal/LockModal'
 import classes from './LockAndViewButtons.module.scss'
 
 const LockAndViewButtons = (): JSX.Element => {
