+++ src/containers/Admin/LogoTab/LogoTab.tsx
@@ -10,8 +10,8 @@ import {
 } from '@entur/typography'
 import { GridContainer, GridItem } from '@entur/grid'
 import { useUser } from '../../../UserProvider'
-import { LoginModal } from '../../../components/LoginModal/LoginModal'
-import { LoginCase } from '../../../components/LoginModal/login-modal-types'
+import { LoginModal } from '../../../components/AccountModals/LoginModal/LoginModal'
+import { LoginCase } from '../../../components/AccountModals/LoginModal/login-modal-types'
 import { LogoUpload } from './LogoUpload/LogoUpload'
 import { SizePicker } from './SizePicker/SizePicker'
 import { Description } from './Description/Description'
