+++ src/containers/Admin/NameTab/NameTab.tsx
@@ -4,8 +4,8 @@ import type { User } from 'firebase/auth'
 import { Heading2, Paragraph } from '@entur/typography'
 import { GridContainer, GridItem } from '@entur/grid'
 import { useUser } from '../../../UserProvider'
-import { LoginModal } from '../../../components/LoginModal/LoginModal'
-import { LoginCase } from '../../../components/LoginModal/login-modal-types'
+import { LoginModal } from '../../../components/AccountModals/LoginModal/LoginModal'
+import { LoginCase } from '../../../components/AccountModals/LoginModal/login-modal-types'
 import './NameTab.scss'
 import { CustomURL } from './CustomURL/CustomURL'
