+++ src/containers/Admin/ShareTab/ShareTab.tsx
@@ -14,10 +14,10 @@ import {
     getOwnersDataByBoardIdAsOwner,
 } from '../../../settings/firebase'
 import { BoardOwnersData, Invite } from '../../../types'
-import { LoginModal } from '../../../components/LoginModal/LoginModal'
-import { LoginCase } from '../../../components/LoginModal/login-modal-types'
-import { RemoveSelfFromTavleModal } from '../../../components/OverflowModals/RemoveSelfFromTavleModal'
-import { NeedToBeOwnerModal } from '../../../components/OverflowModals/NeedToBeOwnerModal'
+import { LoginModal } from '../../../components/AccountModals/LoginModal/LoginModal'
+import { LoginCase } from '../../../components/AccountModals/LoginModal/login-modal-types'
+import { RemoveSelfFromTavleModal } from '../../../components/AccountModals/RemoveSelfFromTavleModal/RemoveSelfFromTavleModal'
+import { NeedToBeOwnerModal } from '../../../components/AccountModals/NeedToBeOwnerModal/NeedToBeOwnerModal'
 import { EditableBoardTitle } from './components/EditableBoardTitle'
 import { BoardLink } from './components/BoardLink'
 import { AddNewOwnersInput } from './components/AddNewOwnersInput'
