+++ src/containers/DashboardWrapper/BottomMenu/buttons/BoardsButton.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { UserIcon } from '@entur/icons'
-import { MineTavlerModal } from '../../../../components/MineTavlerModal/MineTavlerModal'
+import { MineTavlerModal } from '../../../../components/AccountModals/MineTavlerModal/MineTavlerModal'
 import { MenuButton } from '../MenuButton/MenuButton'
 
 const BoardsButton = () => {
