+++ src/containers/Error/ErrorPages.tsx
@@ -3,8 +3,8 @@ import { Helmet } from 'react-helmet'
 import { useNavigate, useParams } from 'react-router-dom'
 import { signOut } from 'firebase/auth'
 import { useToast } from '@entur/alert'
-import { LoginModal } from '../../components/LoginModal/LoginModal'
-import { LoginCase } from '../../components/LoginModal/login-modal-types'
+import { LoginModal } from '../../components/AccountModals/LoginModal/LoginModal'
+import { LoginCase } from '../../components/AccountModals/LoginModal/login-modal-types'
 import { auth, useUser } from '../../UserProvider'
 import { useSettings } from '../../settings/SettingsProvider'
 import sikkerhetBomLight from '../../assets/images/sikkerhet_bom_light@2x.png'
