+++ src/containers/Navbar/Navbar.tsx
@@ -7,8 +7,8 @@ import { GithubIcon, LogOutIcon, PrivacyIcon, UserIcon } from '@entur/icons'
 import { Contrast } from '@entur/layout'
 import { auth, useUser } from '../../UserProvider'
 import { TavlaLogo } from '../../assets/icons'
-import { LoginModal } from '../../components/LoginModal/LoginModal'
-import { LoginCase } from '../../components/LoginModal/login-modal-types'
+import { LoginModal } from '../../components/AccountModals/LoginModal/LoginModal'
+import { LoginCase } from '../../components/AccountModals/LoginModal/login-modal-types'
 import { Theme } from '../../types'
 import classes from './Navbar.module.scss'
